Move request_refund_payment to OffersMessageFlow
